+++ src/Pages/Admin/RightSection/RequestDetails/RequestCardData.json
@@ -1,0 +1,18 @@
+{
+  "results": [
+    {
+      "id": 504685,
+      "onwer": "홍길동",
+      "phone_number": "010-1234-5678",
+      "car_number": "123가4567",
+      "brand": "BMW",
+      "car_name": "M8 그란쿠페",
+      "model_year": "2021",
+      "request_date": "2022.07.10",
+      "name": "BMW 강남 전시장",
+      "dealer": "갑을병",
+      "진행상태": "딜러배정",
+      "진행날짜": "2022.07.10"
+    }
+  ]
+}
