+++ src/Pages/Admin/RightSection/RequestDetails/RequestData.js
@@ -1,0 +1,13 @@
+export const REQUEST_LIST = [
+  { id: 0, title: 'no' },
+  { id: 1, title: '이름' },
+  { id: 2, title: '휴대폰' },
+  { id: 3, title: '차량번호' },
+  { id: 4, title: '브랜드' },
+  { id: 5, title: '모델명' },
+  { id: 6, title: '연식' },
+  { id: 7, title: '견적요청일' },
+  { id: 8, title: '지점' },
+  { id: 9, title: '담당자' },
+  { id: 10, title: '진행상태' },
+];
